2021/08/20

+++ src/main/java/life/majiang/community/controller/PublishController.java
@@ -1,4 +1,5 @@
 package life.majiang.community.controller;
+
 import life.majiang.community.dto.QuestionDTO;
 import life.majiang.community.model.Question;
 import life.majiang.community.model.User;
@@ -19,52 +20,53 @@ public class PublishController {
     private QuestionService questionService;
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id") Integer id,
-                       Model model){
-        QuestionDTO question=questionService.getById(id);
-        model.addAttribute("title",question.getTitle());
-        model.addAttribute("description",question.getDescription());
-        model.addAttribute("tag",question.getTag());
-        model.addAttribute("id",question.getId());
+    public String edit(@PathVariable(name = "id") Long id,
+                       Model model) {
+        QuestionDTO question = questionService.getById(id);
+        model.addAttribute("title", question.getTitle());
+        model.addAttribute("description", question.getDescription());
+        model.addAttribute("tag", question.getTag());
+        model.addAttribute("id", question.getId());
         return "publish";
     }
+
     @GetMapping("/publish")
-    public String publish(){
+    public String publish() {
         return "publish";
     }
 
     @PostMapping("/publish")
-    public String doPublish(@RequestParam(value ="title",required = false) String title,
-                            @RequestParam(value ="description",required = false) String description,
-                            @RequestParam(value ="tag",required = false) String tag,
-                            @RequestParam(value ="id" ,required = false) Integer id,
+    public String doPublish(@RequestParam(value = "title", required = false) String title,
+                            @RequestParam(value = "description", required = false) String description,
+                            @RequestParam(value = "tag", required = false) String tag,
+                            @RequestParam(value = "id", required = false) Long id,
                             HttpServletRequest request,
-                            Model model){
-        model.addAttribute("title",title);
-        model.addAttribute("description",description);
-        model.addAttribute("tag",tag);
+                            Model model) {
+        model.addAttribute("title", title);
+        model.addAttribute("description", description);
+        model.addAttribute("tag", tag);
 
-        if (title==null||title==""){
-            model.addAttribute("error","标题不能为空");
+        if (title == null || title == "") {
+            model.addAttribute("error", "标题不能为空");
             return "publish";
         }
-        if (description==null||description==""){
-            model.addAttribute("error","问题补充不能为空");
+        if (description == null || description == "") {
+            model.addAttribute("error", "问题补充不能为空");
             return "publish";
         }
-        if (tag==null||tag==""){
-            model.addAttribute("error","标签 不能为空");
+        if (tag == null || tag == "") {
+            model.addAttribute("error", "标签 不能为空");
             return "publish";
         }
 
 
-        User user=(User) request.getSession().getAttribute("user");
+        User user = (User) request.getSession().getAttribute("user");
 
-        if (user==null){
-            model.addAttribute("error","用户未登录");
+        if (user == null) {
+            model.addAttribute("error", "用户未登录");
             return "publish";
         }
-        Question question=new Question();
+        Question question = new Question();
         question.setTitle(title);
         question.setDescription(description);
         question.setTag(tag);
