+++ src/main/java/life/majiang/community/controller/QuestionController.java
@@ -1,4 +1,5 @@
 package life.majiang.community.controller;
+
 import life.majiang.community.dto.QuestionDTO;
 import life.majiang.community.service.QuestionService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,12 +15,12 @@ public class QuestionController {
     private QuestionService questionService;
 
     @GetMapping("/question/{id}")
-    public String question(@PathVariable(name = "id") Integer id, Model model){
-        QuestionDTO questionDTO=questionService.getById(id);
+    public String question(@PathVariable(name = "id") Long id, Model model) {
+        QuestionDTO questionDTO = questionService.getById(id);
         //累加阅读量
         questionService.incView(id);
 
-        model.addAttribute("question",questionDTO);
+        model.addAttribute("question", questionDTO);
         return "question";
     }
 }
